Add /surveys endpoint
This commit adds a /surveys endpoint providing a mechanism for the
grouping of related questionnaires into a logical grouping. A
questionnaire now can't be created via the /schema endpoint without
creating a survey first. Once a survey is created with a given
survey_id, that survey_id can be used in requests to the /schema
endpoint.